Move JS to own file; prettify HTML
It's probably not a good idea to have JS in the HTML like we do now, so
I created a new file called countdown.js for it.

I also noticed we had an extra closing section tag on what was line 153.
I removed this line, which allowed my HMTL formatter to clean up the
code a little bit.